Core/Vehicles: add a way to delay the Vehicle despawn Time if its needed.
Port From (https://github.com/TrinityCore/TrinityCore/commit/c7b10d3c5fd082faf6c2835239e340babc3dfc76)

+++ Source/Game/Entities/Player/Player.cs
@@ -232,7 +232,7 @@ namespace Game.Entities
             SetInventorySlotCount(InventorySlots.DefaultSize);
 
             // set starting level
-            SetLevel((uint)GetStartLevel(createInfo.RaceId, createInfo.ClassId, createInfo.TemplateSet));
+            SetLevel(GetStartLevel(createInfo.RaceId, createInfo.ClassId, createInfo.TemplateSet));
 
             InitRunes();
